LayoutTests/imported/w3c/web-platform-tests/css/css-values/attr-length-specified-expected.txt


FAIL type(<length>) lowercase assert_equals: expected "3em" but got ""
FAIL type(<length>) uppercase assert_equals: expected "3em" but got ""
FAIL Unit (em) number assert_equals: expected "3em" but got ""
